+++ packages/emp-frameworks/emp-remote-md5/index.js
@@ -1,0 +1,24 @@
+const Axios = require('axios')
+const md5 = require('md5')
+
+const empRemoteMd5 = async config => {
+  const remotesKey = Object.keys(config.remotes)
+  for (const remoteKey of remotesKey) {
+    const remote = config.remotes[remoteKey].split('@')
+    const fileUrl = remote[remote.length - 1]
+    const response = await Axios({
+      url: fileUrl,
+      method: 'GET',
+    })
+    const fileMd5 = md5(response.data)
+    console.log(fileMd5)
+    if (remote.length > 0) {
+      config.remotes[remoteKey] = `${remote[0]}@${fileUrl}?md5=${fileMd5}`
+    } else {
+      config.remotes[remoteKey] = fileUrl
+    }
+  }
+  return config
+}
+
+module.exports = empRemoteMd5
