+++ projects/demo1/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@efox/emp-plugin-chameleon-topic": "^1.0.0",
+    "@efox/emp-remote-md5": "^1.0.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
   },
